+++ android/src/main/kotlin/io/scer/pdfx/document/Page.kt
@@ -1,13 +1,10 @@
 package io.scer.pdfx.document
 
-import android.annotation.TargetApi
 import android.graphics.Bitmap
 import android.graphics.pdf.PdfRenderer
-import android.os.Build
 import io.scer.pdfx.utils.toFile
 import java.io.File
 
-@TargetApi(Build.VERSION_CODES.LOLLIPOP)
 class Page(
     val id: String,
     private val documentId: String,