+++ android/src/main/kotlin/io/scer/pdfx/resources/DocumentRepository.kt
@@ -1,13 +1,10 @@
 package io.scer.pdfx.resources
 
-import android.annotation.TargetApi
 import android.graphics.pdf.PdfRenderer
-import android.os.Build
 import android.os.ParcelFileDescriptor
 import io.scer.pdfx.document.Document
 import io.scer.pdfx.utils.randomID
 
-@TargetApi(Build.VERSION_CODES.LOLLIPOP)
 class DocumentRepository : Repository<Document>() {
     /**
      * Register document in repository