+++ android/src/main/kotlin/io/scer/pdfx/resources/PageRepository.kt
@@ -1,12 +1,9 @@
 package io.scer.pdfx.resources
 
-import android.annotation.TargetApi
 import android.graphics.pdf.PdfRenderer
-import android.os.Build
 import io.scer.pdfx.document.Page
 import io.scer.pdfx.utils.randomID
 
-@TargetApi(Build.VERSION_CODES.LOLLIPOP)
 class PageRepository : Repository<Page>() {
     /**
      * Register page in repository